The VTC can set the rigid representation's state to the input pose.
This is useful for the initial update of the VTC when the haptic device
may be far from the initial pose of the rigid representation.
Just letting the spring & damper response occur can cause large
forces/torques and oscillations, limiting the usable range for
stiffness and damping.

An alternative approach would be to ramp up the force & torque.  This
approach seemed simpler, but ramping up would better handle situations
in which the initial input pose would cause a collision.