Fix use of auto in RandomBoxPointGenerator
Windows 32 bit release had a test failure when testing the
RandomBoxPointGenerator. This was due to, what I think, is incorrect
usage of auto. ie:

  auto halfSize = box->getSize() * 0.5;

results in halfsize being an eigen expression, not a Vector3d. This
caused some strange stuff to happen later on when halfSize was used.
This change just specifies halfSize as a Vector3d. And adds a bit more
information to the unit tests for this class if a test fails.