The VTC only sends forces & torques to the output during collisions.
The standard implementation for a virtual tool coupler would be for
the rigid representation to perfectly track the input pose unless
there was a collision.  We can't easily do that based on the way
our physics calculations are setup (we could set the velocity, but
trying that has led to some odd behavior), so here we just disable
the output force unless the collision representation collided with
something in the previous update.

This change gets rid of the drag that is otherwise felt due to the
way the rigid representation lags the input pose during free motion.