Fix non-const reference in DataGroupCopier
The DataGroupCopier was taking a non-const reference in its constructor.
And then holding on to the refernce as a member variable. Probably not a
great idea. This changes the non-const reference to a pointer, but
doesn't hold on to it. You need to re-pass the source and target to the
copy, which is a bit safer. However, it does mean that you need to be
sure the IndexDirectories of the source/target passed to the constructor
is the same as the ones passed to the copy. This is a bit better then
before, where the copy needed the inputData reference it was passed
during construction to be the same object as the one passed during
handleInput. Which it is in practice, but the interface suggests it
could be different (it passes inputData to handleInput).