OctreeDcdContact no longer adds contacts to temporary collision reps.
OctreeDcdContact::calculateContactWithNodes uses a temporary collision
pair to check for collisions by delegating to another
DcdContact class which calls CollisionPair::addContact.
Then if the node is a leaf, calculateContactWithNodes transfers any
contacts to a non-temporary CollisionPair.

During the delegation, any addContact call will alter the non-octree's
representation, which ends up with contacts that are not reciprocated
by the octree.  This commit clears those extraneous contacts.