Make id optional in yaml serialization
Make the id optional for components in a yaml file to make authoring
yaml files easier. The id is only used for references in the yaml file,
ie if one component referes to another. The id isn't used for the
constructed component in the simulation anyways. In fact a new id is
generated when the component is constructed. So this change just makes
it easier to write yaml files, you don't need to include an id when it
isn't needed.