The VTC directly sets the velocity of the rigid rep in free motion.
This is similar to the way canonical VTCs work, but it causes an issue
when collisions were happening but are not anymore (like if the rigid
rep rotates out of a collision).  So we zero the velocities once there
is a collision to prevent instabilities.

This requires a low value for the MLCP solver's contact tolerance, like
2e-7.