Increase sleeps in BasicThreadTests
The PauseResumeUpdateTest was failing every once in a while due to the
thread not having enough time to run it's doUpdate. In all other
BasicThread tests we are waiting at least 100ms for similar situations.
In this test, we were only giving the thread 10ms to get through its
doUpdate. This change just gives the thread 100ms to get through its
doUpdate in PauseResumeUpdateTest. And hopefully, fix the failing tests.