Adds VelocityFromPose, a DeviceFilter to estimate velocity from poses.
Most posed devices provide a pose but no velocity.  Many calculations
benefit from having a velocity estimate (e.g., the VirtualToolCoupler).

This commit adds a DeviceFilter that estimates the linear and angular
velocities from the input poses over multiple updates. To do so, it
uses two KalmanFilters, one for linear and one for angular. Each has
a six-vector for the state.
For the linear velocity estimate we use x,v for the state, and a system
in which the velocity is assumed to stay constant between measurements.
For the angular velocity measurement we use w,alpha and a system in
which the acceleration is assumed to stay constant.  I researched Kalman
filtering to estimate angular velocity from orientation measurements
but all the workable solutions used nonlinear Kalman filters.

We may want to add the ability to set the measurement & process noise
magnitudes.  Currently the values were hand-selected for one setup.

Also, some devices may benefit from knowing their own velocities, e.g.,
in NovintScaffold we make velocity-based calculations, so we might
want to replace the DeviceFilter approach with another API,
e.g., each device could hold an instance of a wrapper around a
KalmanFilter.

Another possible issue is that the KalmanFilter is theoretically for
measurements with a constant update period, and here we're always
re-calculating the update period, but it works and a DeviceFilter
doesn't know what update rate it's going to see.