Fix the FindOptiTrack.cmake
Because the name pass to find_package_handle_standard_args was all caps,
it wasn't correctly giving an error when the package wasn't found. This
just renames it so it has the correct capitalization, and now gives an
error if the includes, root direcotory, or libraries are not found (and
the package is required).