NovintDevice can use serial numbers instead of initialization names.
This way an application can ignore the config file and have an
application-specific way of handling multiple Novint devices.

The config file has been renamed devices.yaml and restructured to
have a Novint section.  In the future it may be handled by
CommonDevice.

Multiple devices can be initialized without specifying either the
initialization name or serial number, in which case the next
available device is used.