NovintScaffold bugfixes for new hdl call order with SharedInstance.
Two changes are necessary for the SharedInstance to properly destruct
when all the devices unregister.
The first is that the
NovintScaffold::Handle has to have a weak_ptr to the NovintScaffold,
presumably because now the scaffold holds onto a State that has a
vector of Handles and so if the Handles each have a shared_ptr to the
Scaffold then the Scaffold won't destruct.
The second is that when a device is unregistered, if it was the last
registered device, then we finalize the SDK (destroying the Callback
and all the Handles).  I tried using a weak_ptr<NovintScaffold> in the
Callback, but that caused crashes.
I'm not sure why that second change is necessary, but without it
the NovintScaffoldTest fails because the Scaffold isn't destructing.

This also cleans up the Handle::create and Handle::destroy to meet our
style of single return statement.

Also cleans up the unit tests.