Adds Glut and Boost to the VisualTest target_link_libraries.
GLUT and Boost::thread are used by the VisualTest files, so they
should be linked to VisualTestCommon, so that all the various device
VisualTests don't need to separately link to those dependencies.

Also cleans up the devices' VisualTest CmakeLists, linking to what is
used directly by the tests.

With this, I can build all the visual tests but for Phantom and TrackIR.
PhantomDevice has unresolved externals.
TrackIRDevice is failing to find BitPackedPrecisionMode.

...TrackIRDevice doesn't link to its hardware libraries, and copies the
DLLs around.