Improves VelocityFromPose performance with better process noise.
The process noise is a measure of the increase in noise when the
filter predicts the new state, and is a function of the time since the
last update.  Also, since the new position is a function
of the new velocity, noise in the velocity causes noise in the
position.  Preferably the shape would be calculated from turning
the continuous noise estimate into a discrete estimate.
We may want to consider a more rigorous calculation/estimation for the
process noise, but this change is giving better velocity estimates.

Also reduced the measurement noise and increased the process noise.
More effort tweaking/investigating those parameters would probably be
good.