Make TriangleMeshes Assets
Make TriangleMeshs Assets so they can handle their own loading, this
moves the functionality in the loadTriangleMesh into the doLoad of the
TriangleMesh, and derived class if it has specific data to load as well.
As a result the MeshUtilities and TriangleMeshUtilites are removed since
they aren't needed anymore. This makes the TriangleMesh more like the
OctreeNode.

Make MeshShape and SurfaceMeshShape derive from TriangleMesh, instead of
having a TriangleMesh. This makes sense, and gets rid of the
getMesh()->getMesh calls. Also, simplifies loading, as it is the same as
a regular TriangleMesh. As a result of this, what was TriangleMesh (a
TriangleMeshBase with NormalData) and MeshShape have merged. Also,
TriangleMeshBase has been renamed back to TriangleMesh.

To not have to duplicate changes to the MeshShape, the SurfaceMeshShape
now derives from MeshShape, and only re-implements
computeVolumeIntegrals. Which was originally the only difference between
MeshShape and SurfaceMeshShape anyways.

Finally, MeshRepresentation and OsgMeshRepresentation can now take a
Mesh as an Asset. To fall in line with how we use Assets in OSS, ie
adding a loadMesh, setMesh calls.