Change a feature in OsgView.cpp
When enabling a camera manipulator on an OsgView,
osgViewer::View::setCameraManipulator() has a 2nd parameter, specifying
whether or not to use the home position of the manipulator.
The parameter is defaulted to 'true', which will always use the home
position of the manipulator.

This feature might not be what is wanted: sometimes, a user will set the
manipulator position and manipulator look at on the manipulator first,
then enable the manipulator on an osgViewer::View.
If the 2nd parameter (mentioned above) is 'true' (default value), the
home position of the manipulator will always be used. And this may cause
some confusions.

This commit overwrites the default value to 'false', which will use
whatever values stored in the manipulator itself instead of the home
position.