Attempt to make boosts sleep_until more accurate
Instead of sleep_until for the entire period, do a sleep_until up to
some threshold before the desired time point, then do a busy wait. This
allow low update rate BasicThread to sleep most of the time, but still
hit their desired rate. For high update threads (like Physics) this
would cause it to busy wait most of the time. This is desired, since
sleep_until is not accurate enough for high update rate threads.