Refactors NovintScaffold to push all the inputs before pulling outputs.
This approach lets the InputConsumerInterface modify the output data
based on the current inputs, before that output data gets pulled
by the scaffold and used to render forces on the hardware device.
Previously, the output was pulled before the new pose was read,
so the forces were always (at least) one update behind the pose.

Also, removes two members of the DeviceData that temporarily held
the pose and orientation before it was put into scaledPose. Instead,
scaledPose itself is used in those places.

Also, checkDeviceHoming now *only* checks the homing, and doesn't
change/zero the pose.