Fix windows build of the graphics mesh
Fix the windows build of SurgSim::Graphics::Mesh where it didn't like
the static definition of m_className in the SurgSim::Graphics namespace.

Also, figured out that the original changes to
SURGSIM_ADD_SERIALIZABLE_PROPERTY are not needed. What was really going
on is that the compilier couldn't convert a Mesh to an Asset because the
Mesh was forward declared. Added the Mesh.h include to fix this.

Finally, just switched the Mesh used in the SerializationTest to the
Cube, since it is in the OsgMeshRepresentationTests directory.