Adds a MockInputComponent.
For Behaviors that interact with devices, the unit tests often
benefit from being able to get known data from the InputComponent.
This class exposes the DataGroup, so it can be set in the test.
That means that no DeviceInterface is required, and so no
InputManager is needed (which typically connects the InputComponent
and DeviceInterface).