Minor fixes to Meshes
Replaced (triangle.isValid == true) with (triangle.isValid) for
simplicity.

Used using for dependant name resolution of doClearVertices in the
TriangleMesh, instead of using this->

Remove unneeded default constructor for Vertex, as it isn't needed and
leaves the position uninitialized.

Fold MeshShape::copyWithTransform into MeshShape::setPose, since it is
small, and that is the only place it is used.

Cleaned up some incorrect variable names in MeshShapeTests.