Override MassSpringRepresentation::doInitialize => add Spring::initialize
This commit synchronizes Spring and FemElement a bit more. As we always try to keep them in sync as much as possible.

The last difference is FemElement::update() which was introduced to handle the corotational model recalculation of the elements rotation.
But this is wrong as called only once per time step (FemElement::update() is called in the physics computation afterUpdate).
This calculation and update of the matrix for FemElement corotational needs to happen each time we call computeF/M/D/K/FMDK, so the method FemElement::update() will go away.

+ Factorized the LinearSpring matrix calculation for clarity.
+ Got rid of all addSubMatrix, addSubVector, getSubVector, replaced by direct Eigen access (block and segment).
+ Check assertion failure in the unit test