Fix Fem3DElementCorotationalTet to recompute its matrices in addF/M/D/K/FMDK
This element was using update() to recompute the rotation and update the matrices, but this was happening only in afterUpdate() which happens once per time step.
So an implicit OdeSolver iterating to compute the solution with precision requiring a re-evaluation of the system for different state would not have gotten what it wanted.
It would have gotten the same matrices at each iteration, ruining the effort of the non-linear solver to find a precise solution.

+ This change gets rid of the update() method in FemElement as it was its only purpose, giving the occasion to the Corotational model to update the rotation and the matrices
+ This change gets rid of FemRepresentation::afterUpdate as it was its only purpose, to call update on the FemElements.
+ Refactor Fem3DElementCorotationalTet matrices calculation for clarity.
+ Change the UnitTest accordingly.