Fix when the BasicThread timing is logged
There was an incorrect comparison when checking to see if the
BasicThread timing info should be logged. It checked if the current
logging threshold was greater than or equal to LOG_LEVEL_INFO. Meaning
that if the log level threshold was set to SEVERE, it would enter the
block. But not log anything if the log level was lower than INFO, like
DEBUG. This change just corrects this.

Also, simplified that logging output for the actual and requested rate.