Create protected PhysicsManager::addComputations
Instead of making the private m_computations protected, for derived
classes to use, create an protected addComputations, for derived classes
to add computations, and not be able to mess with m_computations.

Also, remove initializeComputations. The PhysicsManager already has a
protected virtual doInitialize. All this function was doing was calling
initializeComputations. So instead, derived classes should just follow
the pattern of our other classes, and override doInitialize to
initialize the PhysicsManager differently (ie, adding different
computations).