Removes VelocityFromPose DeviceFilter.
This class is not currently used and we no longer expect to need
it anytime soon.  Since its unit tests are failing on the build
server (probably due to timing issues since the class uses an
internal timer), this commit removes it.  If we need it in the
future, we'll re-add it.