CollisionReps go inactive if MeshShape fails to update or setPose.
MeshShape::doUpdate and MeshShape::setPose both call calculateNormals
which can calculate an indeterminate normal if a triangle has
coincident vertices.  This commit makes the collision rep go inactive
in that situation.

One issue that having an indefinite normal can cause is for
TriangleTriangleIntersection to assert because it cannot find
the intersection with another triangle during
TriangleMeshTriangleMeshDcdContact.