Fix deadlock in ThreadPool desctruct
Fix a bug where the m_destructing was set after a thread checked it
and before it starts to wait, and also misses the notify_all. By
locking before setting m_destructing to true, we make sure the threads
are either already waiting, or running a task and will see the boolean
when they are done.

Also, handle spurious wakeups of the wait, which can happen, but rarely.
Here we add the predicate to the wait command. This basically checks if
the wait was intented to be woken up (when there are tasks, or shutdown
is occuring).

Add doxygen tparam to ThreadPool::enqueue.

Finally, clean up the ThreadLoop to be a little more clear. I think.