The Stapling example uses a specialized PhysicsManager.
It sets the max number of MLCP solve iterations to 200...I've
observed successful solutions that took 130 iterations.

It also does not have the UpdateCollisionRepresentation computation
that the base PhysicsManager does after PushResults, which reduces
the time spent updating AABB trees for the MeshShapes.  The
collision representations will be stale after PushResults until
following FreeMotion, but that won't affect the simulation.

This increases the update rate during contact from ~130Hz to ~150Hz.