Stapling example, the InputComponent is posed.
This commit removes the poses that were placed on the arm and wound
to bring them to a good position relative to the device.
Now than InputComponents and OutputComponents can be posed, the
correct way to position a scene relative to the device is to set
the local pose on the InputComponents and OutputComponents.

Also, brings the camera closer to the wound.