Stapling example has a DeviceFactory so it can use other devices.
You must still build MultiAxisDevice to build the Stapling example.
The 7Dof Falcon is not working because we need to roll-back the
changes to NovintScaffold.

The change to the PhantomScaffold removes the linear velocity
input data.  That data was never significantly tested and it seems
to be making the VTC unstable.