Limit the number of staples in the scene
When the number of stales reach 10 in the scene, the program will not
allow the user to add more staple, unless they reset the staples.

This is to prevent the program to crash because of too many staples will
probably put too many constraints (over-constraint) the FEM.
However, if multiple staples create the same constraints on the FEM,
things will be fine.