Fix a couple of GCC security warnings
Ubuntu turns on extra GCC security warnings by default in GCC. This
addresses these issues, the first adds checking of the return value
from fread. The second removes the use of variable format parameter to
fscanf. Not using a string literal as the format in fscanf is a
security vulnerability.