Can getData from OutputComponent, for data passing between Behaviors.
Sometimes one Behavior depends on the output data from another Behavior.
For example, a metric for how much force the simulation is exerting
on the user via a haptic device through the VirtualToolCoupler.
The VTC calculates the force and sets it to an OutputComponent, which
stores the data in its member OutputProducer.  Any DeviceInterface
that has been connected to the OutputComponent will then call
requestOutput directly on the OutputProducer. DeviceInterface and
Behavior do not share a base class, so we need another approach for a
Behavior (e.g., a metric) to get the data from an OutputComponent.