Moved posed shapes to the Collision::Representation
Previously, the mesh shape internally kept an initial mesh, and had a
setPose that would modify the mesh w.r.t. this initial mesh. This change
moves this shape caching into the Collision Representation, where it is
actually being used.  Collision Representation has a "posed shape" that
is the shape, transformed by the pose of the representation. This shape
is cached, so subsequent calls to getPosedShape don't incure a cost if
the pose is the same as the last call. The original mesh shape is now in
an un-posed state, which is consistent with the other shapes.

This change is also applied to ParticleShapes that are similar to meshes
in that it derives from Vertices, and can also benefit from caching.