SparseMatrix: Investigatory branch for MLCP timing
This commit contains a umber of intermediate code changes
to allow us o explore timing issues in the Stapling demo
following the introduction of the sparse matrix modifications.
It restores the buffered calculation of the compliance matrix
and the matrix x sparse vector multiply at the heart of the
MLCP calculation.