Use Vertices for Particles
This simplifies Particles and makes them more consistent with
OpenSurgSim by using the Vertices structure. In the process the
Particle, ParticleReference, and ParticlesState data is now stored in
the new vertices structure (Particles) directly. Also, storing used and
unused particles in two std::list was actually found to be slower than
adding and removing particles from a std::vector, so a boost in
performance is actually seen in this simplification.