Fix clearing of collisions with one representation
If there is only one collision representation, its collisions were not
being cleared. Granted, this would only be rarely be an issue, unless
something collided with something else, and then one of them became
inactive. This clears the collisions correctly, and puts in the same
function that publishes the collisions for clarity.