Add ParticleCollisionResponse Computation
Add a physics computation that handles collision response for Particle
Representations. This computation calls
Particles::Representations::handleCollisions which calls an abstract
private function doHandleCollisions that is implemented for a specific
particle system implementation. This commit adds the implementation for
SphRepresentation.

In addition, have DcdCollision calculate collisions for Particle
Representations by adding TriangleMeshParticlesDcdContact to the
DcdCollision table.

Update the RenderTestSphRepresentation to demonstrate collisions between
particles and a mesh using this new computation.