Fix bug preventing usage of new FemElement (switch an assert to an if statement)
We set automatically the ElementType depending on what the Asset contains.
But we the element is not recognized by OSS, it probably means that the Element
used is defined in a module, so the code should not assert but only set the
element type automatically if it recognize it, do nothing otherwise. In the
latter case, the user should call setFemElementType manually (before or after
setFem).