Fix tracking mode may not being reported correctly
To enable tracking mode, "Auto-orient Tracking" must be off in the
LeapControlPanel. If it is on, HMD mode can't be set, the Leap will just
detect what to do. If this is the case, getTrackingMode would report the
wrong value, as it doesn't actually check to see if the policy got
enacted, only reports back the value set. This change actually checks
that the policy was set, so getTrackingMode gives the correct value.