Adds log messages at start and end of Runtime::addSceneElements.
This is helpful for debugging issues with the YAML files because
when an Accessible is being decoded we don't get any information about
problems except possibly the name of the property.  This way we can
at least narrow it down to a single yaml file.

Also, fixes a bug in Accessible::decode.  When the property is
not found, a log message is printed, but if the value being set
is not a Scalar type, trying to convert it to a string throws
an exception and the log message doesn't get printed. So this commit
first checks if the value is a Scalar.