Make Collision::Rep contact registration thread safe
When DcdCollision becomes multi-threaded, multiple contacts can be added to the
same Collision::Rep at the same time, causing raise condition in accessing the
map<shared_ptr<Rep>, list<Contact>>.

This commit makes the write-access to this data structure thread-safe.
We rely on the fact that the read-access will only happen after DcdCollision is
done, at which point the data structure will be completed and fixed.