Set cmake policy CMPL0054 to NEW.
Cmake began warning about this policy on version 3.2.2.

"Only interpret if() arguments as variables or keywords when unquoted."
For example, we do:
if("${CMAKE_CXX_COMPILER_ID}" MATCHES "Clang")
and we do not want to treat the "Clang" as a variable because the
cmake compiler ids are supposed to be those specific strings.