Accessible::decode logs the node to DEBUG.
When decoding nodes, if setting a property causes some sort of
issue inside the instance being decoded, it's very useful to know
about the node being decoded.  E.g., if GenerateTangents is true
for an osg file that does not have vertices then TangentSpaceGenerator
logs a warning but that doesn't tell me which node is the issue.