If LabJackScaffold::configureDevice fails, initialize returns false.
Previously LabJackThread::doInitialize would assert.

configureDevice needs to be called by LabJackThread because it needs
to be called by the same thread that will call
LabJackScaffold::runInputFrame.
Previously configureDevice was being called by
LabJackThread::doInitialize, which meant that when the configuration
failed (e.g., because a USB re-enumeration was in progress) BasicThread
would assert.  Now, configureDevice is called in LabJackThread's
constructor.  The thread is created in LabJackScaffold::registerDevice
which then checks to see if the configuration succeeded, and if not
it returns false which causes LabJackDevice::initialize to return false.

This approach fits our typical model wherein a failure for a device
to setup/configure should return false from initialize() and let
the caller decide how to respond.