Retry opening communications with LabJack.
We have been observing situations in which the LabJack fails to open,
then opens fine after we restart our program.  We are going to try
opening communications a few times, with a delay in between each try.

This will make the SEARCH options for model and connection slow because
each combination that fails will cause a few seconds of delay.