InputVtc links to SURGSIM_DEVICES_LIBS instead of hard-coded libraries.
If we use the new ObjectFactory in DeviceInterface, then each device
library that gets built adds itself to the factory, so we need to
link to all the built device libraries.

In this commit, the Devices CMakeLists generates a list of the
libraries and passes that information up to its parent, which passes
it up to the top level, so all Examples and Modules have access to
the variable.  Since SurgSim is configured before Examples or Modules,
the variable is always ready before we would use it.