Move Deformable initial state transformation from doWakeUp to doInitialize
This is a bug in OSS as someone could request the initial state from the deformableRep after adding it to the scene, but before the Runtime runs the simulation (behavior, analysis...) in which case, it would get invalid information as the state has not been transformed yet.
This requires the setLocalPose to be called before adding the Component into the Scene, but makes the construction process more robust.