Packaging: Gives finer control over the modules to be packaged.
This commit gives finer control over the packaging process. Every
package aware module should assign their install components to a
named COMPONENT in the install command and should publish a boolean
to the CMakeCache named CPACK_PACKAGE_<ModuleName>. If the boolean
is on, the upper level module CMakeList file should also set the
CPACK_PACKAGE_NAME and the CPACK_COMPONENTS_ALL variables
appropriately to include their modules in the package and pass the
variables up to the next higher level. For example, the following
form can be used:

if (${CPACK_PACKAGE_BURRHOLE})
       set(CPACK_PACKAGE_NAME "${CPACK_PACKAGE_NAME}<ModuleName>"
       		PARENT_SCOPE)
       set(CPACK_COMPONENTS_ALL ${CPACK_COMPONENTS_ALL}
          	<ModuleComponents> PARENT_SCOPE)
endif (${CPACK_PACKAGE_BURRHOLE})

This will bring a boolean out to the CMakeCache, that if set will
cause <ModuleComponents> to be included in any packager output.
Multiple modules and the OSS developers toolset can be packaged
either together or separately using this model.

There are two caveats:
	1 - Setting all the install booleans off for a packaging
	operation will install all named COMPONENTS along with the
	any "Unspecified" components. This is the default case.
	2 - Whether a component is selected for packaging or not
	does not affect the behavior of the INSTALL target.