Move createDevice to Devices.h put device libraries in SurgSimDevices
Previously, when using createDevice, you had to ensure to include the h
file for each device you wanted to use (for static registration and
factory registration of the device). This could be done also if you
remembered to include Devices.h. In addition, you had to be sure to
link to the device library using a cmake variable
SURGSIM_DEVICE_LIBRARIES. This change put createDevice into a free
function that is included in Devices.h. This make it easier to not
forget to include this, since if you include Devices.h you get all the
built device headers (and the devices in the factors) as well as
createDevice. In addition, all the build devices are linked against a
new library (SurgSimDevices). So to use all the avaliable devices, just
link against this. So now using createDevice is a bit simpler. include
Devices.h, and link to SurgSimDevices.

Also, clarified the createDevice documentation, and switched the
parameter order to match ObjectFactory::create. ie, className first,
than parameter name. It gets a bit confusing if you are using the device
factory and the createDevice function.