Don't assert if Oculus not connected
The oculus asserts if it can't initialize the sdk (and connect to the
deamon). This prevents the controlling program from having a chance to
handle this and try a substitude device. This moves the sdk
initialization into doInitialize, and returns false if it fails.

Also, remove the __FUNCTION__ from the loging messages, we don't
normalling do this.