DeviceInterface needs a virtual destructor.
We expect to hold derived classes (devices) in DeviceInterface,
a base class pointer, and we need the derived classes' destructors
to be called because they cleanup the device state.
A future commit may move all the cleanup into the finalize function,
and then the DeviceInterface destructor can just call finalize
and there won't be a need for the derived classes to define destructors.