Add properties to LeadDevice
This adds Accessible properties to the LeapDevice. To do this, the enum
for hand type and tracking mode are removed, and these settings are set
directly with functions. These functions can now be easily hooked up as
properties and used without having to include LeapDevice.h (for the
enums).

Also, adds unit tests for creation through a factory, and using the
properties.