Adds back enum based hand type setting to LeapDevice
To make the interface clearer, add back the enum based hand type setting
to the LeapDevice. Ie, keep it to one function that can set the hand
type to left or right handed tracking.

But also adds a way to set the hand type by string to address the issue
of setting the hand type with an Accessible property without having to
inlcude LeapDevice.h (for the enum).

Also, adds a conveniance boost::any convert function from const char *
to std::string. This makes it easier to call setValue on a property that
is a std::string (ie, you can pass in a string literal).

Finally, fixes the getClassName documentation in DeviceInterface and the
default getClassName returned by CommonDevice.