Refactors NovintScaffold.  Fixes circular dependency.
Also, more canonical use of the HDAL which fixes crashes in the
unit tests.

The circular dependency was because the Callback held onto a shared_ptr
to the scaffold instance.  It was supposed to be fixed by calling
finalizeSdk, but there was a bug if registerDevice failed that kept
the scaffold alive for the rest of the program.

In this commit, neither Handle nor Callback hold onto
the scaffold instance.  To enable that change, checkForFatalError is
now a free function in an unnamed namespace, NovintScaffold::getLogger
has been removed, and the Callback takes a raw pointer to the scaffold.

To make sure that the HDAL is initialized and un-initialized correctly,
the calls are made in NovintScaffold's constructor and destructor.
Also, the registeredDevices list is not cleared until after the
callback is stopped, so that the destructor does not need to lock
the state's mutex (which caused blocking runHapticFrame).
These two changes seems to have fixed the crashes we've been seeing in
the unit tests.

Also, adds many Debug log messages to the scaffold. These have been
helpful in straightening out when hdal functions are being called.