Cleanup LabJackScaffold. Removes unnecessary circular reference.
Removes getLogger from the scaffold.  getLogger tends to lead to
circular references with DeviceData and it isn't necessary since
we can just get the logger from the Framework. That means the
Handle doesn't need to hold onto a pointer back to the scaffold.

Refactors the Handle::create function into the constructor since
that's the only place the code should be run.

Refactors the Thread destruction so that it's performed in the
DeviceData destructor, which removes specialized code in
registerDevice and unregisterDevice.  DeviceData destructor also
handles calling Handle::destroy.