DeviceInterface declares isInitialized, and finalize is now public.
finalize() is the opposite of initialize() and it's useful for the
calling code to be able to un-initialize a device (particularly when
multiple devices/DeviceFilters are connected and the combined device's
initialization fails).

This also updates initialize() for MultiAxisDevice and FilteredDevice,
so that all sub-devices are either initialized or have been finalized,
not some in one state and some in the other.

isInitialized is useful because finalize should only be called on
initialized devices.  Many devices already defined isInitialized.