Adds DeviceFilter, a base class for device filters.
The DeviceFilter is a concrete class that will simply pass data
through unchanged, but its functions can be overridden to perform
filtering on input and/or output.

Removes datagroup indices storage from the filters because it
complicates the code and testing has shown that it does not
significantly improve performance.