Adds CollisionTypes and OnlyCollideWith to Collision Representation
Adds the ability to specify a Collision::Representation will only
collide with a set of other Collision::Representations (similar to
ignore).

Adds a CollisionType and SelfCollisionType for each
Collision::Representation. This type can either be discrete, continuous,
or none. A CollisionType is also added to the CollisionPair, and is
determined based on what the CollisionType's of the representations in
the pair.

Modifies DcdCollision to only perform collision detection on
CollisionPairs that have a discrete CollisionType.