Reworks CollisionTypes and OnlyCollideWith
Renames OnlyCollideWith to Allow for setting which collision
representation is only allowed to collide with.

Renames the COLLISON_TYPE enum back to COLLISON_DETECTION_TYPE for
clarity. All associated functions using this enum are also modified.

Reverts back to using addDcdContact and addCcdContact on the
CollisionPair, but then immediatly asserts if the CollisionPair is
itself a different type (DCD/CCD).

Adds unit test for the SURGSIM_SERIALIZABLE_ENUM.