Removes SparseMatrix blockOperation for Sparse sub matrices
This removes the SparseMatrix blockOperation overload for performing
block operations on a SparseMatrix with a SparseMatrix sub matrix. This
functionality is not used in our code currently. All uses of the
blockOperation on a SparseMatrix use a dense sub matrix. The reason for
this change is that Eigen 3.3 moved the InnerIterator typedef out of
SparseMatrixBase into a new sub-class of SpraseMatrixBase,
SparseCompressedBase. If we moved the blockOperation to accept
SparseCompressedBase, the code would no longer be compatible with Eigen
3.2, nor would it be able to accept Eigen::Block<Eigen::SparseMatrix>,
which is subclassed from SparseMatrixBase. Since we currently don't use
this function, I think removing it is the simplist way forward.

Also, the Linear Sparse solver is higher precision in Eigen 3.3. This
removes a test the that solver is not high precision. This is a good
thing.