Compresses a sparse matrix for SparseMatrix tests
On debian-sid-i386 a unit test was failing because one of the
SparseMatrices was not compressed before it was used. This actually
caused an exception to not be thrown when using the matrix, when it
should have thrown an exceptioiin. For some reason this did not show up
on the other architectures.