Disables long running Truth cube tests
The truth cube tests have gotten out of date. They do not use the
current physics pipeline for performing the test, and does all the
calculations manually in the test. As a result, it does not actualy
tests the physics ability to correctly simulate the truth cube data. In
addition, the test is taking hours to complete on some platforms.
Therefor, this change disables the tests until they can be updated to
correctly test physics.